+++ src/app/app.routes.ts
@@ -1,3 +1,23 @@
 import { Routes } from '@angular/router';
+import { authGuard } from './auth/auth.guard';
 
-export const routes: Routes = [];
+export const routes: Routes = [
+  {
+    path: 'login',
+    loadComponent: () => import('./login/login').then((m) => m.Login),
+  },
+  {
+    path: 'board',
+    loadComponent: () => import('./kanban/kanban').then((m) => m.Kanban),
+    canActivate: [authGuard],
+  },
+  {
+    path: '',
+    redirectTo: 'board',
+    pathMatch: 'full',
+  },
+  {
+    path: '**',
+    redirectTo: 'board',
+  },
+];